// libs/list-types/interim-applications-chd-daily-cause-list/src/models/types.ts
export interface InterimApplicationHearing {
  judge: string;
  time: string;
  venue: string;
  type: string;
  caseNumber: string;
  caseName: string;
  additionalInformation: string;
}

export interface OpenJusticeStatementDetail {
  nameToBeDisplayed: string;
  email: string;
}

export interface InterimApplicationsChdData {
  hearingList: InterimApplicationHearing[];
  openJusticeStatementDetails: OpenJusticeStatementDetail[];
}